pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fspipeline"
version = "0.1.0"
description = "Few-shot TTS training data extraction pipeline for CosyVoice3"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "ffmpeg-python>=0.2.0",
    "torch>=2.1",
    "torchaudio>=2.1",
    "pyannote.audio>=3.1",
    "speechbrain>=1.0",
    "faster-whisper>=1.0",
    "transformers>=4.40",
    "openai>=1.12",
    "soundfile>=0.12",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.3",
]

[project.scripts]
fspipeline = "fspipeline.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
